Update patch set 2
Patch Set 2:

It's not clear that it would be worth it -- it's never pulling any significant amount of data down, and it also happens during "repo gitc-init", which doesn't take a -j option currently (and those options are parsed in two places, so it's not easy to update).

So I could plumb it through from "repo sync", but as long as -j32 seems like a reasonable number from the server point of view, I think this is probably good for now. We don't support gitc outside of the googlesource.com, once we did, this answer would likely change.

Patch-set: 2