Update patch set 3
Patch Set 3:

To further: You could not crop the URL in this way when the URL ends with /ssh_info or /Gerrit (both of which are cropped from the URL already).

This would work directly with what I have, and just requires verbosity in the other direction. However it may break other peoples systems so while it's clean it may not be a best choice.

Patch-set: 3