Update patch set 2
Patch Set 2:

Dan, is it possible to have a look to this patch again? Currently, repo sync with depth 1 fails if a project revision is fixed to sha1. This patch allows to workaround this case by fetching all branches (or branch specified in upstream). It allows to make fetch with depth=1 work for all other projects with revision set to branch or tag. I think this is better than failure.

Patch-set: 2