Update patch set 2
Patch Set 2:

(1 comment)

> Currently, repo sync with depth 1 fails if a project revision is fixed to sha1.

This statement isn't true for us -- I don't know whether that's normal, or something special with our server, but we're using this successfully.

We have manifests that use clone-depth="1" and revision="sha1..." where this change would cause severe problems (there isn't an upstream marked, so it would download all branches). In my test, for a single project in one of those manifests, this increased the download required from ~300MB to ~10GB.

There's code in _RemoteFetch to retry the download without a depth if we failed to download the SHA1 we were looking for -- is that not working?

Which version of git are you using? Are you able to fetch a non-tip shallow clone from this server?

  mkdir test; cd test; git init .
  git remote add origin https://gerrit.googlesource.com/git-repo
  git fetch --depth=1 origin 490d09a31415d3fd1b16f650188bfd8e701ae8e8

Patch-set: 2
Label: Verified=0