Update patch set 1
Patch Set 1: Code-Review+1

Note: This is a first of a series of changes to make repo work on Windows with the native version of python (as opposed to the cygwin version of windows).

The end goal is to add the ability to run "repo" from a native windows command prompt, using the native Windows port of Windows and the "Git For Windows" version of git, hence removing the dependency on cygwin.

Note that repo should still work with Cygwin even with these changes -- the platform_utils.isWindows() checks returns "true" only when running with the native windows port of python.

Note also that I am trying to make sure that the non-windows code paths remain exactly the same (hence the "isWindows()" checks), as opposed to trying to use portable code everywhere.

Patch-set: 1
Label: Code-Review=+1