Update patch set 1
Patch Set 1:

I did the original investigation a while ago and can't remember the details but I believe on Windows the output is piped through a pager and at that point isatty starts returning False.

I'll have another look and see if I can give a better answer to that question. I agree that it's not clear that this code is strictly more correct than the old version.

Patch-set: 1