init: add --submodules to sync manifest submodules
repo sync can sync submodules via the --fetch-submodules option. However, if the
manifest repo has submodules, those will not be synced. Having submodules in the
manifest repo -- while not commonly done -- can be useful for inheriting a
manifest from another project using <include> and layering changes on top of it.
In this way, you can avoid having to deal with merge conflicts between your own
manifests and the other project's manifests (for example, if you're managing an
Android fork).

Add a --submodule option to init that automatically syncs the submodules in the
manifest repo whenever the manifest repo changes.

Change-Id: I45d34f04517774c1462d7f233f482d1d81a332a8