Update patch set 1
Patch Set 1:

(10 comments)

the schema should be documented, and the commit message should provide some background information

i was thinking repo would produce simple machine output that could be ingested into higher tools (like chromite), but i guess having it write a dedicated format directly wouldn't be a bad idea.  if that is the route you want to take, i'd suggest trying to create a centralized module (instead of stuffing a random subcommand) and making --event-log a `repo` command rather than a subcommand.  that way you have the infra to add eventpoints and such to other places of interest in repo.

Patch-set: 1