Update patch set 1
Patch Set 1:

(1 comment)

the sync format still needs to be documented in this tree ... repo can't depend on other projects (like chromite)

wrt things worth enumerating, in the CrOS world, the big ones are probably the two halves of syncing (network & local).  but once you have infra for enumerating, seems like you could just sprinkle high level "start" & "finish" events for the subcommands themselves in the main repo logic.  then we'd be able to do a run, ingest all the data, and then know if there are other repo commands that are doing a lot of work we didn't know about :).

Patch-set: 1