fix prune for tags used as manifest revision
When definiting the revision of a project in manifest file, one
can define a tag using 'refs/tags/tagname'. This works fine for
init, sync and upload, but will fail for prune.

During prune operation, repo tries to reset HEAD to the git hash
of the tag 'refs/tags/tagname', which is an invalid target for HEAD.
Instead, repo should really do rev_parse to get the valid
commit hash, that the tag corresponds to.

Change-Id: I387f4aeea4e2e2b82bcaa2858cc1543a6ac90e85