Add documentation on parsing order
Some tools that have been created to parse manifest.xml files appear to
vary from the parsing order that manifest_xml.py uses. Codify its
behavior in documentation so that the other tools can be corrected and
authorative documentation can be referenced.

One such problem seen is that the "remove-project" nodes are applied
before all the "project" nodes in a file. This causes the following
example XML to fail:

  <?xml version="1.0" encoding="UTF-8"?>
  <manifest>
    <!-- Main manifest projects grouped separately to avoid merge
         conflicts. -->
    <project path="foo" name="upstream/foo" />
    <project path="bar" name="upstream/bar" />
    <project path="baz" name="upstream/baz" />

    <!-- Add all modifications to main manifest here to avoid merge
         conflicts. -->
    <remove-project name="bar" />
    <project path="bar" name="next-gen/bar" />
  </manifest>

Change-Id: I507cacff3fd9bc9a00be4b04509cf5ed0e9527eb