Add documentation on parsing order
Some tools that have been created to parse manifest.xml files appear to
vary from the parsing order that manifest_xml.py uses. Codify its
behavior in documentation so that the other tools can be corrected and
authorative documentation can be referenced.

One such problem seen in a third-party tool is that the "remove-project"
nodes are not applied. This behavior causes the "bar" project to not be
removed in the following XML:

  <?xml version="1.0" encoding="UTF-8"?>
  <manifest>
    <!-- Main manifest projects grouped separately to avoid merge
         conflicts. -->
    <project path="foo" name="upstream/foo" />
    <project path="bar" name="upstream/bar" />
    <project path="baz" name="upstream/baz" />

    <!-- Add all modifications to main manifest here to avoid merge
         conflicts. -->
    <remove-project name="bar" />
  </manifest>

Change-Id: I507cacff3fd9bc9a00be4b04509cf5ed0e9527eb