Fix the initial existence check for "repo"
The _Clone function does not actually clone in the snese of calling "git
clone" but it only initializes and fetches, i.e. the checkout is
missing. Hence the existence check for "repo" must be moved to after the
checkout has happened.

While at it, also slightly improve the code by using os.path.join().

Change-Id: Ied89e24231addabab6075005065748df1ffa74c4