Add "post-sync" repo hook
This hook is executed after sync is run. A dictionary of changed revisions and
two lists of deleted and added projects are passed.

This hook is just a notification mechanism and can't affect the outcome of the
"repo sync" command. It runs afterwards only if the "repo sync" command
succeeds.

A simple post-sync hook implementation:

def main(project_revs, added_projects, deleted_projects, **kwargs):
  for project in added_projects:
    print ('added project: {}'.format (project))

  for project in deleted_projects:
    print ('deleted project: {}'.format (project))

  for project, revs in project_revs.items():
    print ('project: {}, old: {}, new: {}'.format (project, revs.old, revs.new))

Change-Id: Iaaf925244a1ee0967fd1e13e3f8508a11b073c59