Update patch set 6
Patch Set 6: Code-Review-2

I think this is adding a hook to avoid addressing a real problem, and I think that's a recipe for technical debt that we won't be able to handle.

That said, I do think there's a real problem you're trying to solve, and happy to help with that (either here or on repo-discuss@).

Patch-set: 6
Reviewer: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-2