Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> i don't mind having a post-sync hook (in CrOS we every once in a while want a way to "clean up" or "reset" certain things after a sync), but i'm not fully understanding the use case here or what a post-sync hook would even do.  what does git bisect have to do with `repo sync` ?  or differentiating between OE-specific git trees and git trees coming from other vendors ?
> 
> maybe it would help if you described what your post-sync hook would actually be doing ?

We have a project that contains a dependencies manifest that we symlink as a local manifest. This almost work as desired for us but there are some things left to polish.

When we run "repo sync" the local (symlinked) manifest may change, but when it does "repo" won't see those changes on the same "repo sync" call, as the local manifest is read by repo before the fetching of the project (containing it as a symlink) happens. Because of this, additional "repo sync <dependency project>" calls are required when the symlinked local manifest changes, otherwise the dependencies are not updated until the next "repo sync" call and things are left in an inconsistent state in between. There is a 1 "repo sync" call delay for the symlinked local manifest.

Now we are solving (workarounding) this problem with git hooks that diff the local manifest (symlinked), so when it changes they just invoke "repo sync <dependencies>" to bring the dependencies up to date, avoiding that "repo sync" 1 call delay.

Solving (workarounding) this with git hooks is more complicated than it would be doing it on a "post-sync" hook, as there are more hooks to maintain and there are unavoidable false positive "sync" operations when doing local git operations on the project. We did it with git hooks as a workaround.

Sane in-project bisecting/checkout/branching is unrelated to the hooks. It's made possible because we are symlinking our dependencies manifest as a local manifest. As on typical OE setups that use repo all the dependencies are in the manifest at the top level, we can't just do "git checkouts/bisects, etc" on our project/layer and get our dependencies updated to the revision that they were in that point of time, for that we need backing of the manifest itself too (which isn't possible without external infrastructure, as there is no implicit "project revision" to "manifest revision" link). Having the local manifest symlinked and contained on our project allows us to "git checkout + repo sync" and to be able to build directly.

I hope that this solves your questions.

IMO, the use that we would make of this hook shouldn't affect the outcome of this review, as having a "post-sync" hook as a notification mechanism is a small, sane, low-impact, useful feature by itself and the time to write it is already spent. If we shoot ourselves in our foot by doing a bad use of it wouldn't be because of a repo tool inconsistency.

So in short we are bringing in submanifests from a project and allowing people to use our layer without bringing in a full manifest with all our dependencies.

Submanifests would be a nice repo feature, but provided that it's sane (which maybe isn't) I'm definitely not able to contribute it both by time budget, knowledge and ability to test it.

Patch-set: 6
Reviewer: Gerrit User 1071136 <1071136@173816e5-2b9a-37c3-8a2e-48639d4f1153>