Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-2
> 
> I think this is adding a hook to avoid addressing a real problem, and I think that's a recipe for technical debt that we won't be able to handle.
> 
> That said, I do think there's a real problem you're trying to solve, and happy to help with that (either here or on repo-discuss@).

This must be some misunderstanding. 

This is an external contribution, I'm not working at Google so the usage of this isn't introducing (hypothetical) technical debt for you or anyone in your organization to handle but for me/us. 

From your POV I'm just contributing an external patch which IMO is sane, small and can be useful for many people out (we are just three people on this review and Mike and I have uses for it). A simple and innocent "post-sync" notification hook.

"git" has hooks for every operation, I don't see why "repo" should just have only one hook ("pre-upload") selected on arbitrary grounds from the software design POV. So why deny free external hook contributions on Google if they are that small, sane and low-impact?

From my POV you are -2 ing this because of the use that I said that we would do (outside of Google) not on the sanity/correctness/usefulness of the feature on technical grounds. As I said, this must be some misunderstanding.

Patch-set: 6
Reviewer: Gerrit User 1071136 <1071136@173816e5-2b9a-37c3-8a2e-48639d4f1153>