Update patch set 6
Patch Set 6:

> This is an external contribution, I'm not working at Google so the usage of this isn't introducing (hypothetical) technical debt for you or anyone in your organization to handle but for me/us. 

I'm sorry, but this is not how open source works. I'm speaking as someone who helps with the repo project, not as a Googler. Every change has costs and benefits, so I'm trying to better understand the benefits and figure out a way to bring them about with minimal cost.

Sometimes a more invasive change costs less in the long run. So, for example, for this use case it might make sense for repo to learn some concept of recursive manifests (or to use submodules for the sub-sub-projects or something) instead of this small hook.

> > someone think it's not safe.
> 
> https://gerrit-review.googlesource.com/c/git-repo/+/70874

That's from before https://gerrit-review.googlesource.com/c/git-repo/+/83051, which clarified the security model.

Patch-set: 6
Reviewer: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>