Update patch set 7
Patch Set 7:

> You are judging something as generic as a repo hook considering only one of the possible use cases. Mike Frysinger, k cai and Aymen Bouaziz seem to have different use cases for it that your review doesn't cover. You are -2 ing their use cases too then.

I'm trying to understand the use case in order to make sure that the patch we accept addresses it well.

I still don't understand the use case. If vapier says that the resulting patch handles the (quite distinct) case he is interested in, then that's great, and I have nothing against that. Likewise if the case you're describing starts making sense to me.

What I really don't want to do is to make this overly generic or to promise to continue to support a use case that isn't workable, or to add a hook *without* already knowing what the use case behind it is. Hopefully that clarifies my objection.

You might wonder why I think it's my business to ask what a feature is for. But it really does matter. I have learned from experience that it costs a lot not to ask.

Patch-set: 7
Reviewer: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>