Update patch set 2
Patch Set 2: Code-Review+2

Not about this change: at some point it would be nice to switch to push options instead of % syntax. That said, very old installations of Gerrit may not support push options yet.

Should we look for characters that are not allowed in a refname and switch to push options when any are present?

Patch-set: 2
Reviewer: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2