Add option to filter projects down to those that have local branches.
This is useful to speed up "repo status" and various scripts, using the
assumption that a) the user starts at least one local branch in each
project they work in, and b) they work on a small number of projects.

This is based on the observation that I always use "repo branch" rather
than "repo status" just because it's so much faster and typically lists
all I want to see anyway.

Change-Id: Ia4ab4dbbd7be1ee7a8a9a4559edd81f26346fbeb