Update patch set 1
Patch Set 1:

> Is there a way to make the code explicitly call out conflicts (i.e. to explicitly say "I want -c here to override -c from init" while having other conflicts still be diagnosed)?

only on an ad-hoc basis.  gitc-init could detect it after the fact in _Options and remove things, or it could pass a flag to the init _Options to tell it to not add -c.

but imo, it's better we keep the interface stable.  if we train people to use init's options, and then train them that gitc-init supports everything init does, except for when there's conflicts and then the options might mean something completely different (like -c), that's bad UI imo.

Patch-set: 1