Update patch set 2
Patch Set 2: Verified+1

so the code & the comments didn't seem to line up.  the generator docstring says:

> A unique, to this invocation of the program, integer id.

but the way _EventIdGenerator is written & used means it wasn't actually unique to the program, it was unique to the instantiation of the generator which is unique to the instantiation of EventLog.  in practice, i think EventLog only gets instantiated once, so we ended up with unique ids across invocation of the program.

PS1 changed the id generation to match the code behavior (unique-per-instantiation), but since i think we really want per-program-invocation, PS2 went that route by using a module-level multiprocessing.Value.

Patch-set: 2
Reviewer: Gerrit User 1007145 <1007145@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1