color: Allow strings with no formatting args to contain "%"
If "repo init" was run in a path containing "%", "repo info" would fail
printing the path with

    File ".repo/repo/color.py", line 173, in f
      return fmt % args
    TypeError: not enough arguments for format string

as the "%" in the path name is interpreted as the start of a formatting
specifier. As paths are output without any formatting args, work around
that error by not even trying to interpolate the string if no args are
specified and thus printing the literal "%" in the path as-is.

Change-Id: I1ea36ec9518c21938b3c75b263306d98951e704b