Update patch set 1
Patch Set 1:

this isn't really an upload, this is a direct push.  the --help text is extremely terse and easy for people to mess up.

is this something that we need a dedicated flag for ?  this is such an uncommon scenario, is it problematic to have people do `git push` for these one-offs instead ?

Patch-set: 1
Reviewer: Gerrit User 1007145 <1007145@173816e5-2b9a-37c3-8a2e-48639d4f1153>