sync: Add --no-checkout option
In some cases it may be desirable to provision the actual repository
working trees from an alternate source. (eg: from a shared development
image) in lieu of having to always sync the source tree, especially in
large projects.  This change introduces a '--no-checkout' option to
the 'repo sync' CLI that enables us to configure the general state of
the git repository without actually having to update the working
tree.  This involves (1) configuring the underlying .git directory as
required, (2) setting up the HEAD reference correctly.  The intended
approach is intended to provide a similar behaviour to 'git clone
--no-checkout'.

For subsequent 'repo sync' invocations we may be using
Project#_Checkout() to update the working tree state. If we don't want
to checkout a new working tree (eg: we've specified repo sync
--no-checkout) we want to use reset instead of checkout so that the
working tree is left intact.

Additionally, when we're merging changes into the local branch we
can't touch the working tree, instead we want to reset the branch and
working tree state.

Change-Id: I298b2e78bb0bb30481d56024ca260ab435b67d0b