Update patch set 3
Patch Set 3:

(2 comments)

> Out of curiosity - I am having issues with reviewing my Gerrit submissions in AOSP too. What is the proper way to speed the review up? Adding commit authors as reviewers does not help.

it largely depends on the specific repo in question and how active people are.  unfortunately it's not uncommon in AOSP for people to really only be active on the Google/internal side of things.

you could take a look at the OWNERS file in the repo if there are any.  the people listed in there are supposed to be responsible for reviewing.

Patch-set: 3
Reviewer: Gerrit User 1007145 <1007145@173816e5-2b9a-37c3-8a2e-48639d4f1153>