repo: force use of Python 3
We don't change the shebang because program name usage is spotty at
best: on some platforms (like macOS), it's not uncommon to not have
a `python3` wrapper, only a major.minor one like `python3.6`.  Using
python3 wouldn't guarantee a new enough version of Python 3 anyways,
and we don't want to require Python 3.6 exactly, just that minimum.

So we check the current Python version.  If it's Python 2, we exec
`python3`.  If that fails, we fallback to searching for a newer
`python3.X` version.  If it's already Python 3, but not new enough,
we fallback to searching `python3.X` too.

If we still can't bootstrap up to a new enough version, we just die.

Bug: https://crbug.com/gerrit/10418
Change-Id: I5e157defc788e31efb3e21e93f53fabdc7d75a3c