Update patch set 2
Patch Set 2:

(1 comment)

this CL is only half the solution.  this runs only at manifest parsing time and the only context it has is the manifest itself.  the other half (i have pending locally) is executed after sync has finished and just before the actual copy/link funcs run.  at that point we'd have full access to the filesystem and be able to pull off any fun tricks like you've suggested.

i was thinking of splitting the two so that the filesystem half wouldn't have to deal with such completely pathological inputs.  it's already going to be complicated, so i'm trying to keep a lid on that, especially since it's important to not get it wrong :).  the manifest check is able to filter out a lot of things that we know we'd never want, and do so quickly & before we sync any git trees (although, all things considered, failing early here isn't super important as we're talking about an error-only situation).  it does make the unittests run fast at least.

i'll see about finishing up the 2nd half that needs to touch the filesystem and upload it.  that'll give us something concrete to consider.

Patch-set: 2
Reviewer: Gerrit User 1007145 <1007145@173816e5-2b9a-37c3-8a2e-48639d4f1153>